Organization, Optimizations
Moved method that determines penalty for grass to the car class.
Tweaked the friction factor of grass. Changed use of Point2D class to a
custom “Point” class, so that there is better space management. Updated
the boundary of the car so that it is the tightest boundary possible
(not estimate). Removed clutter. Changed the README to reflect updates.